+++ backend/src/modules/items/items.module.ts
@@ -4,6 +4,7 @@ import { TypeOrmModule } from "@nestjs/typeorm";
 import { ItemCode } from "./entities/item-code.entity";
 import { Item } from "./entities/item.entity";
 import { StockLevel } from "../stock/entities/stock-level.entity";
+import { StockMovement } from "../stock/entities/stock-movement.entity";
 import { Location } from "../locations/entities/location.entity";
 import { Supplier } from "../suppliers/entities/supplier.entity";
 import { ItemsController } from "./items.controller";
@@ -12,7 +13,7 @@ import { AccessControlModule } from "../access-control/access-control.module";
 import { LoggingModule } from "../logging/logging.module";
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Item, ItemCode, StockLevel, Location, Supplier]), AccessControlModule, LoggingModule],
+  imports: [TypeOrmModule.forFeature([Item, ItemCode, StockLevel, StockMovement, Location, Supplier]), AccessControlModule, LoggingModule],
   controllers: [ItemsController],
   providers: [ItemsService],
   exports: [ItemsService],